% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runDataApp}
\alias{runDataApp}
\title{Launch test shiny Apps}
\usage{
runDataApp(name = "usarrest", ...)
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch test shiny Apps
}
\examples{
\dontrun{
library(usarrestApp)

# running an app
  runDataApp()
}

}
