% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runDataApp}
\alias{runDataApp}
\title{Launch test shiny Apps}
\usage{
runDataApp(name = "usarrest", ...)
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch test shiny Apps
}
\examples{
\dontrun{
library(usarrestApp)

# running an app
  runDataApp()

# to get list of names of shiny apps
  runDataApp()
# this will give error with all possible
# shiny names contained in the package.
}

}
